wip(model): add map of hexagons
